pyproject.toml
[tool.poetry]
name = "flask-htmx-test"
version = "0.1.0"
description = ""
authors = ["beret <beret@hipisi.org.pl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-fasthtml = "^0.4.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
